Taulak egokituta (aldagai aldaketa) eta motak zuzenduta (7)

+++ models/biki_lagina.py
@@ -8,11 +8,11 @@ class Lagina(models.Model):
     _description = 'Laginak gordetzeko'
 
     NumMuestra = fields.Integer(index=True,required=True)
-    PosicionReal = fields.Integer()
-    VelocidadReal = fields.Integer()
-    VelocidadConsigna = fields.Integer()
-    CargaReal = fields.Integer()
-    CargaContacto = fields.Integer()
+    PosicionReal = fields.Float()
+    VelocidadReal = fields.Float()
+    VelocidadConsigna = fields.Float()
+    CargaReal = fields.Float()
+    CargaContacto = fields.Float()
     Egoera = fields.Many2one('biki.egoera', string='Egoera', index=True,required=True)
     Exekuzioa = fields.Many2one('biki.exekuzioa', string='Exekuzioa', index=True,required=True)
